+++ packages/api/lib/controllers/boxesController.js
@@ -381,9 +381,10 @@ const getBox = async function getBox (req, res, next) {
  * @apiParam (RequestBody) {Location} location the coordinates of this senseBox.
  * @apiParam (RequestBody) {String="homeV2Lora","homeV2Ethernet","homeV2Wifi","homeEthernet","homeWifi","homeEthernetFeinstaub","homeWifiFeinstaub","luftdaten_sds011","luftdaten_sds011_dht11","luftdaten_sds011_dht22","luftdaten_sds011_bmp180","luftdaten_sds011_bme280","hackair_home_v2"} [model] specify the model if you want to use a predefined senseBox model, autocreating sensor definitions.
  * @apiParam (RequestBody) {Sensor[]} [sensors] an array containing the sensors of this senseBox. Only use if `model` is unspecified.
- * @apiParam (RequestBody) {String="hdc1080","bmp280","tsl45315","veml6070","sds011","bme680","smt50","soundlevelmeter", "windspeed"} [sensorTemplates] Specify which sensors should be included.
+ * @apiParam (RequestBody) {String="hdc1080","bmp280","tsl45315","veml6070","sds011","bme680","smt50","soundlevelmeter", "windspeed", "scd30"} [sensorTemplates] Specify which sensors should be included.
  * @apiParam (RequestBody) {Object} [mqtt] specify parameters of the MQTT integration for external measurement upload. Please see below for the accepted parameters
  * @apiParam (RequestBody) {Object} [ttn] specify parameters for the TTN integration for measurement from TheThingsNetwork.org upload. Please see below for the accepted parameters
+ * @apiParam (RequestBody) {Boolean="true","false"} [useAuth] whether to use access_token or not for authentication
  *
  * @apiUse LocationBody
  * @apiUse SensorBody
@@ -408,15 +409,16 @@ const postNewBox = async function postNewBox (req, res, next) {
  * @api {get} /boxes/:senseBoxId/script Download the Arduino script for your senseBox
  * @apiName getSketch
  * @apiGroup Boxes
- * @apiParam {String="Serial1","Serial2"} serialPort the serial port the SDS011 sensor is connected to
- * @apiParam {String="A","B","C"} soilDigitalPort the digital port the SMT50 sensor is connected to
- * @apiParam {String="A","B","C"} soundMeterPort the digital port the soundlevelmeter sensor is connected to
- * @apiParam {String="A","B","C"} windSpeedPort the digital port the windspeed sensor is connected to
- * @apiParam {String} ssid the ssid of your wifi network
- * @apiParam {String} password the password of your wifi network
- * @apiParam {String} devEUI the devEUI of TTN device
- * @apiParam {String} appEUI the appEUI of TTN application
- * @apiParam {String} appKey the appKey of TTN application
+ * @apiParam {String="Serial1","Serial2"} [serialPort] the serial port the SDS011 sensor is connected to
+ * @apiParam {String="A","B","C"} [soilDigitalPort] the digital port the SMT50 sensor is connected to
+ * @apiParam {String="A","B","C"} [soundMeterPort] the digital port the soundlevelmeter sensor is connected to
+ * @apiParam {String="A","B","C"} [windSpeedPort] the digital port the windspeed sensor is connected to
+ * @apiParam {String} [ssid] the ssid of your wifi network
+ * @apiParam {String} [password] the password of your wifi network
+ * @apiParam {String} [devEUI] the devEUI of TTN device
+ * @apiParam {String} [appEUI] the appEUI of TTN application
+ * @apiParam {String} [appKey] the appKey of TTN application
+ * @apiParam {Boolean="true","false"} [display_enabled] include code for an attached oled display
  * @apiUse JWTokenAuth
  * @apiUse BoxIdParam
  */
@@ -424,7 +426,8 @@ const getSketch = async function getSketch (req, res, next) {
   res.header('Content-Type', 'text/plain; charset=utf-8');
   try {
     const box = await Box.findBoxById(req._userParams.boxId, { populate: false, lean: false });
-    res.send(box.getSketch({
+
+    const params = {
       serialPort: req._userParams.serialPort,
       soilDigitalPort: req._userParams.soilDigitalPort,
       soundMeterPort: req._userParams.soundMeterPort,
@@ -433,8 +436,16 @@ const getSketch = async function getSketch (req, res, next) {
       password: req._userParams.password,
       devEUI: req._userParams.devEUI,
       appEUI: req._userParams.appEUI,
-      appKey: req._userParams.appKey
-    }));
+      appKey: req._userParams.appKey,
+      display_enabled: req._userParams.display_enabled
+    };
+
+    // pass access token only if useAuth is true and access_token is available
+    if (box.access_token) {
+      params.access_token = box.access_token;
+    }
+
+    res.send(box.getSketch(params));
   } catch (err) {
     handleError(err, next);
   }
@@ -442,7 +453,7 @@ const getSketch = async function getSketch (req, res, next) {
 
 /**
  * @api {delete} /boxes/:senseBoxId Mark a senseBox and its measurements for deletion
- * @apiDescription This will delete all the measurements of the senseBox. Please not that the deletion isn't happening immediately.
+ * @apiDescription This will delete all the measurements of the senseBox. Please note that the deletion isn't happening immediately.
  * @apiName deleteBox
  * @apiGroup Boxes
  * @apiUse ContentTypeJSON
@@ -488,6 +499,7 @@ module.exports = {
       { name: 'devEUI', dataType: 'StringWithEmpty' },
       { name: 'appEUI', dataType: 'StringWithEmpty' },
       { name: 'appKey', dataType: 'StringWithEmpty' },
+      { name: 'display_enabled', allowedValues: ['true', 'false'] },
     ]),
     checkPrivilege,
     getSketch
@@ -506,7 +518,9 @@ module.exports = {
       { name: 'ttn', dataType: 'object' },
       { name: 'sensors', dataType: ['object'] },
       { name: 'addons', dataType: 'object' },
-      { predef: 'location' }
+      { predef: 'location' },
+      { name: 'useAuth', allowedValues: ['true', 'false'] },
+      { name: 'generate_access_token', allowedValues: ['true', 'false'] }
     ]),
     checkPrivilege,
     updateBox
@@ -530,13 +544,14 @@ module.exports = {
       { name: 'exposure', allowedValues: Box.BOX_VALID_EXPOSURES },
       { name: 'model', allowedValues: Box.BOX_VALID_MODELS },
       { name: 'sensors', dataType: ['object'] },
-      { name: 'sensorTemplates', dataType: ['String'], allowedValues: ['hdc1080', 'bmp280', 'sds 011', 'tsl45315', 'veml6070', 'bme680', 'smt50', 'soundlevelmeter', 'windspeed'] },
+      { name: 'sensorTemplates', dataType: ['String'], allowedValues: ['hdc1080', 'bmp280', 'sds 011', 'tsl45315', 'veml6070', 'bme680', 'smt50', 'soundlevelmeter', 'windspeed', 'scd30'] },
       { name: 'serialPort', dataType: 'String', defaultValue: 'Serial1', allowedValues: ['Serial1', 'Serial2'] },
       { name: 'soilDigitalPort', dataType: 'String', defaultValue: 'A', allowedValues: ['A', 'B', 'C'] },
       { name: 'soundMeterPort', dataType: 'String', defaultValue: 'B', allowedValues: ['A', 'B', 'C'] },
       { name: 'windSpeedPort', dataType: 'String', defaultValue: 'C', allowedValues: ['A', 'B', 'C'] },
       { name: 'mqtt', dataType: 'object' },
       { name: 'ttn', dataType: 'object' },
+      { name: 'useAuth', allowedValues: ['true', 'false'] },
       { predef: 'location', required: true }
     ]),
     postNewBox