+++ packages/api/lib/controllers/measurementsController.js
@@ -11,7 +11,7 @@ const
   handleError = require('../helpers/errorHandler'),
   OutlierTransformer = require('../transformers/outlierTransformer'),
   jsonstringify = require('stringify-stream'),
-  { UnauthorizedError } = require('restify-errors');
+  { UnauthorizedError, NotFoundError } = require('restify-errors');
 
 /**
  * @api {get} /boxes/:senseBoxId/sensors Get latest measurements of a senseBox
@@ -20,11 +20,49 @@ const
  * @apiName getLatestMeasurements
  * @apiUse BoxIdParam
  */
+/**
+ * @api {get} /boxes/:senseBoxId/sensors/:sensorId Get latest measurements of a sensor
+ * @apiDescription Get the latest measurements of a sensor.
+ * @apiGroup Measurements
+ * @apiName getLatestMeasurementOfSensor
+ * @apiUse BoxIdParam
+ * @apiUse SensorIdParam
+ * @apiParam {Boolean="true","false"} [onlyValue] If set to true only returns the measured value without information about the sensor. Requires a sensorId.
+ */
 const getLatestMeasurements = async function getLatestMeasurements (req, res, next) {
+  const { _userParams: params } = req;
+
+  let box;
+
   try {
-    res.send(await Box.findBoxById(req._userParams.boxId, { onlyLastMeasurements: true }));
+    box = await Box.findBoxById(req._userParams.boxId, { onlyLastMeasurements: true });
   } catch (err) {
     handleError(err, next);
+
+    return;
+  }
+
+  if (params.sensorId) {
+    const sensor = box.sensors.find(s => s._id.equals(params.sensorId));
+    if (sensor) {
+      if (params.onlyValue) {
+        if (!sensor.lastMeasurement) {
+          res.send(undefined);
+
+          return;
+        }
+        res.send(sensor.lastMeasurement.value);
+
+        return;
+      }
+      res.send(sensor);
+
+      return;
+    }
+
+    res.send(new NotFoundError(`Sensor with id ${params.sensorId} does not exist`));
+  } else {
+    res.send(box);
   }
 };
 
@@ -65,14 +103,16 @@ const getData = function getData (req, res, next) {
   let stringifier;
 
   // IDEA: add geojson point featurecollection format
-  if (format === 'csv' || (download === 'true')) {
+  if (format === 'csv') {
     res.header('Content-Type', 'text/csv');
-    res.header('Content-Disposition', `attachment; filename=${sensorId}.${format}`);
     stringifier = csvStringifier(['createdAt', 'value'], delimiter);
   } else if (format === 'json') {
     res.header('Content-Type', 'application/json; charset=utf-8');
     // IDEA: add geojson point featurecollection format
     stringifier = jsonstringify({ open: '[', close: ']' }, jsonLocationReplacer);
+  }
+  if (download === 'true') {
+    res.header('Content-Disposition', `attachment; filename=${sensorId}.${format}`);
   }
 
   let measurementsStream = Measurement.getMeasurementsStream(req._userParams)
@@ -181,12 +221,17 @@ const getDataMulti = async function getDataMulti (req, res, next) {
  * @apiParam (RequestBody) {String} value the measured value of the sensor. Also accepts JSON float numbers.
  * @apiParam (RequestBody) {RFC3339Date} [createdAt] the timestamp of the measurement. Should conform to RFC 3339. Is needed when posting with Location Values!
  * @apiParam (RequestBody) {Location} [location] the WGS84-coordinates of the measurement.
+ * @apiHeader {String} Authorization Box' unique access_token. Will be used as authorization token if box has auth enabled (e.g. useAuth: true)
  */
 const postNewMeasurement = async function postNewMeasurement (req, res, next) {
   const { boxId, sensorId, value, createdAt, location } = req._userParams;
 
   try {
     const box = await Box.findBoxById(boxId, { populate: false, lean: false });
+    if (box.useAuth && box.access_token && box.access_token !== req.headers.authorization) {
+      throw new UnauthorizedError('Box access token not valid!');
+    }
+
     const [measurement] = await Measurement.decodeMeasurements([{
       sensor_id: sensorId,
       value,
@@ -234,6 +279,7 @@ const postNewMeasurement = async function postNewMeasurement (req, res, next) {
  * @apiUse LocationBody
  * @apiParam {String} [luftdaten] Specify whatever you want (like `luftdaten=1`. Signals the api to treat the incoming data as luftdaten.info formatted json.
  * * @apiParam {String} [hackair] Specify whatever you want (like `hackair=1`. Signals the api to treat the incoming data as hackair formatted json.
+ * @apiHeader {String} Authorization Box' unique access_token. Will be used as authorization token if box has auth enabled (e.g. useAuth: true)
  * @apiParamExample {application/json} JSON-Object:
  * {
  *   "sensorID": "value",
@@ -290,10 +336,15 @@ const postNewMeasurements = async function postNewMeasurements (req, res, next) 
 
   if (Measurement.hasDecoder(contentType)) {
     try {
-      const box = await Box.findBoxById(boxId, { populate: false, lean: false, projection: { sensors: 1, locations: 1, lastMeasurementAt: 1, currentLocation: 1, model: 1, access_token: 1 } });
+      const box = await Box.findBoxById(boxId, { populate: false, lean: false, projection: { sensors: 1, locations: 1, lastMeasurementAt: 1, currentLocation: 1, model: 1, access_token: 1, useAuth: 1 } });
 
-      if (contentType === 'hackair' && box.access_token !== req.headers.authorization) {
-        throw new UnauthorizedError('Access token not valid!');
+      // if (contentType === 'hackair' && box.access_token !== req.headers.authorization) {
+      //   throw new UnauthorizedError('Box access token not valid!');
+      // }
+
+      // authorization for all boxes that have not opt out
+      if ((box.useAuth || contentType === 'hackair') && box.access_token && box.access_token !== req.headers.authorization) {
+        throw new UnauthorizedError('Box access token not valid!');
       }
 
       const measurements = await Measurement.decodeMeasurements(req.body, { contentType, sensors: box.sensors });
@@ -360,7 +411,9 @@ module.exports = {
   ],
   getLatestMeasurements: [
     retrieveParameters([
-      { predef: 'boxId', required: true }
+      { predef: 'boxId', required: true },
+      { predef: 'sensorId' },
+      { name: 'onlyValue', required: false }
     ]),
     getLatestMeasurements
   ]