+++ packages/models/src/box/sensorLayouts/sensorDefinitions/index.js
@@ -39,7 +39,8 @@ const veml6070_uvintensity = require('./veml6070_uvintensity'),
   smt50_soilmoisture = require('./smt50_soilmoisture'),
   smt50_soiltemperature = require('./smt50_soiltemperature'),
   soundlevelmeter = require('./soundlevelmeter'),
-  windspeed = require('./windspeed');
+  windspeed = require('./windspeed'),
+  scd30_co2 = require('./scd30_co2');
 
 module.exports = {
   hdc1008_temperature,
@@ -81,5 +82,6 @@ module.exports = {
   smt50_soilmoisture,
   smt50_soiltemperature,
   soundlevelmeter,
-  windspeed
+  windspeed,
+  scd30_co2
 };