+++ tests/tests/014-descriptive-statistics-test.js
@@ -25,7 +25,7 @@ describe('openSenseMap API Routes: basic descriptive statistics', function () {
         ]);
       })
       .then(function (responses) {
-        const boxes = responses.map(r => { return { _id: r.body.data._id, sensorid: r.body.data.sensors[0]._id }; });
+        const boxes = responses.map(r => { return { _id: r.body.data._id, sensorid: r.body.data.sensors[0]._id, access_token: r.body.data.access_token }; });
         boxIds = responses.map(r => r.body.data._id).join(',');
 
         for (const box of boxes) {
@@ -108,9 +108,9 @@ ${box.sensorid},3,2018-02-05T14:06:12.620Z
         const [first, second, third] = boxes;
 
         return chakram.all([
-          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes/${first._id}/data`, first.measurements, { json: false, headers: { 'content-type': 'text/csv' } }),
-          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes/${second._id}/data`, second.measurements, { json: false, headers: { 'content-type': 'text/csv' } }),
-          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes/${third._id}/data`, third.measurements, { json: false, headers: { 'content-type': 'text/csv' } }),
+          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes/${first._id}/data`, first.measurements, { json: false, headers: { 'content-type': 'text/csv', 'Authorization': first.access_token } }),
+          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes/${second._id}/data`, second.measurements, { json: false, headers: { 'content-type': 'text/csv', 'Authorization': second.access_token } }),
+          chakram.post(`${process.env.OSEM_TEST_BASE_URL}/boxes/${third._id}/data`, third.measurements, { json: false, headers: { 'content-type': 'text/csv', 'Authorization': third.access_token } }),
         ]);
       });
   });